requirements.txt
Pillow==10.0.0
pillow-heif==0.12.0
tqdm==4.65.0
